Changing reimbursement data collection to conform new version
The brazilian congress made available a new version of the reimbursement
dataset, which now can be downloaded already in csv format and separated
by year. This somewhat simplifies the current fetching routine, but it
also demands a lot of small changes in the way the dataset is aquired,
translated, loaded and merged.

I this rather large commit, the unit tests were also modified to conform
to the new way of fetching, and the "convert to csv" test was removed,
since this operation is not needed anymore. The convert_to_csv() mehtod
from the CEAPDataset class was kept, though, so we don't break rosie
until it removes the call to this method.

Another quite drastic change was made in the test cases' names. A number
was added to them, so we can 'make sure' a proper order is followed
during integration tests. I strongly believe there is a better way of
doing this. I can work on it in a near-future commit.

Signed-off-by: Luiz Carlos Cavalcanti <cavalcanti.luiz@gmail.com>