Change the way we translate subquotas
Since some subquotas appear or disappear over the years,
translating the CEAP yearly files the way we were doing
could cause errors when a subquota was not present in the
file in question. Now we are translating in a 'soft' way,
and also replacing descriptions according to their IDs.

This solves two problems at once: the application does not
break when a subquota is not present and we don't need to
worry about case and punctuation differences.

Signed-off-by: Luiz Carlos Cavalcanti <cavalcanti.luiz@gmail.com>