Introducing Federal Senate to `serenata-toolbox`
This is the first test and script for fetching the Federal Senate datasets.
Soon we will be able to add the new datasets to Amazon and use them normally.

I don't think it needs more cleaning, but I will be studying it later.
The translation and compression tasks are already working, an everything was developed using TDD.
Feel free to help :)