Remove error silencing and specify quotation option
Since the Camara's Open Data team fixed most issues with
CSV formating, we are removing the fail-safe options on
the CSV reading call. Also we are specifying now that
the files have no quotation marks of any kind to separate
values. This last change avoid silent row dropping when
a valeu has no escape characters on quotation for text
values.

Signed-off-by: Luiz Carlos Cavalcanti <cavalcanti.luiz@gmail.com>