Fix datasets availability
This PR changes the datasets from Amazon S3 to Digital Ocean Spaces.
Following that change, a it added a bit of clean up:

* As the idea is to use Digital Ocean Spaces, we have a nice URL
accessbile to the whole team in the Digital Ocean dashboard to
manage datasets files, so this PR completely remove the
`Datasets.remote` object
* Due to the same reason, this PR removes everything related to
authentication (i.e. the toolbox is as read-only tool to the storage
now)

In spite of that it is flexible: when the envvar points to Digital Ocean
Spaces, it works. If this envvar is not set it works with Amazon S3
right away.